requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=1.3.0=pyhd8ed1ab_0
autopep8=1.6.0=pyhd3eb1b0_1
backcall=0.2.0=pypi_0
blas=1.0=mkl
boto3=1.4.5=py36_0
botocore=1.5.92=py36_0
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.12.7=ha878542_0
certifi=2016.2.28=py36_0
click=6.7=py36_0
cloudpickle=0.4.0=py36_0
coverage=4.4.1=py36_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.10.0=py36_0
cytoolz=0.8.2=py36_0
dask-core=0.15.2=py36_0
dataclasses=0.8=pyh787bdff_2
decorator=4.1.2=py36_0
docutils=0.14=py36_0
entrypoints=0.4=pypi_0
ffmpeg=4.3=hf484d3e_0
freetype=2.12.1=hca18f0e_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
grpcio=1.35.0=py36hce63b2e_0
icu=67.1=he1b5a44_0
imageio=2.6.1=py36_0
ipykernel=5.5.6=pypi_0
ipython=7.16.3=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=5.0.0=pypi_0
jbig=2.1=0
jedi=0.17.2=pypi_0
jmespath=0.9.0=py36_0
joblib=0.11=py36_0
jpeg=9e=h0b41bf4_3
jupyter-client=7.1.2=pypi_0
jupyter-core=4.9.2=pypi_0
kiwisolver=1.3.1=py36h2531618_0
lame=3.100=h166bdaf_1003
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=3.0=h295c915_0
libblas=3.9.0=8_mkl
libcblas=3.9.0=8_mkl
libdeflate=1.8=h7f8727e_5
libffi=3.3=h58526e2_2
libgcc-ng=12.2.0=h65d4601_19
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libiconv=1.14=0
liblapack=3.9.0=8_mkl
libpng=1.6.39=h753d276_0
libprotobuf=3.4.0=0
libsqlite=3.40.0=h753d276_0
libstdcxx-ng=12.2.0=h46fd767_19
libtiff=4.3.0=h6f004c6_2
libuv=1.44.2=h166bdaf_0
libwebp-base=1.3.0=h0b41bf4_0
libzlib=1.2.13=h166bdaf_4
llvm-openmp=16.0.1=h417c0b6_0
markdown=2.6.9=py36_0
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py36h5fdd944_1
mkl=2020.4=h726a3e6_304
ncurses=6.3=h27087fc_1
nest-asyncio=1.5.5=pypi_0
nettle=3.6=he412f7d_0
networkx=1.11=py36_0
numpy=1.19.5=py36hfc0c790_2
olefile=0.44=py36_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h0b41bf4_0
pandas=0.25.3=py36hb3f55d8_0
parso=0.7.1=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.2.0=py36h8328e55_2
pip=9.0.1=py36_1
prompt-toolkit=3.0.29=pypi_0
protobuf=3.4.0=py36_0
ptyprocess=0.7.0=pypi_0
pycodestyle=2.10.0=pyhd8ed1ab_0
pygments=2.12.0=pypi_0
pyparsing=2.2.0=py36_0
python=3.6.13=h12debd9_1
python-dateutil=2.6.1=py36_0
python_abi=3.6=2_cp36m
pytorch=1.10.2=py3.6_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2017.2=py36_0
pywavelets=1.1.1=py36h7b6447c_2
pyzmq=22.3.0=pypi_0
readline=8.2=h8228510_1
regex=2017.4.5=py36_0
requests=2.14.2=py36_0
s3transfer=0.1.10=py36_0
sacremoses=0.0.38=py_0
scikit-image=0.15.0=py36hb3f55d8_2
scikit-learn=0.24.2=py36hc89565f_1
scipy=1.5.3=py36h81d768a_1
setuptools=36.4.0=py36_1
six=1.10.0=py36_0
sqlite=3.40.0=h4ff8645_0
tensorboard=1.15.0=py36_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h27826a3_0
toml=0.10.2=pyhd8ed1ab_0
toolz=0.8.2=py36_0
torchaudio=0.10.2=py36_cu113
torchvision=0.11.3=py36_cu113
tornado=4.5.2=py36_0
tqdm=4.15.0=py36_0
traitlets=4.3.3=pypi_0
transformers=2.1.1=py_0
typing_extensions=3.10.0.2=pyha770c72_0
wcwidth=0.2.5=pypi_0
werkzeug=0.12.2=py36_0
wheel=0.29.0=py36_0
widgetsnbextension=4.0.0b0=pypi_0
xz=5.2.6=h166bdaf_0
zlib=1.2.13=h166bdaf_4
zstd=1.5.2=h3eb15da_6
